Added jasmine support. Fixed existing UT to use jasmine
