+++ .eslintrc
@@ -17,6 +17,7 @@
     "brace-style": ["error", "stroustrup"],
     "no-console": "warn",
     "comma-dangle": "off",
-    "no-unused-vars": "off"
+    "no-unused-vars": "off",
+    "no-trailing-spaces": "off"
   }
 }
