+++ package.json
@@ -62,6 +62,7 @@
     "eslint-config-airbnb": "^15.1.0",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-react": "^7.14.3"
+    "eslint-plugin-react": "^7.14.3",
+    "jasmine": "^3.5.0"
   }
 }
